+++ src/lib/useReactDataTable/useReactDataTable.tsx
@@ -1,5 +1,4 @@
-﻿/* eslint-disable max-lines */
-import {
+﻿import {
   getCoreRowModel,
   getExpandedRowModel,
   getFilteredRowModel,
@@ -7,7 +6,7 @@ import {
   getSortedRowModel,
   useReactTable,
 } from "@tanstack/react-table";
-import { useReactDataTableState, useReactDataTableStateProps } from "../useReactDataTableState/useReactDataTableState";
+import { useReactDataTableState } from "../useReactDataTableState/useReactDataTableState";
 import Skeleton from "react-loading-skeleton";
 import { useReactDataTableProps } from "./useReactDataTableProps";
 import { useReactDataTableResult } from "./useReactDataTableResult";
@@ -18,10 +17,14 @@ import { getSortingStateFromModel } from "../utils/getSortingStateFromModel";
 import { getModelFromSortingState } from "../utils/getModelFromSortingState";
 import { OptionalNullable } from "../types/NullableTypes";
 import { useMemo } from "react";
+import { useReactDataTableStateProps } from "../useReactDataTableState/useReactDataTableStateProps";
 
 /**
  * A react hook that returns a react table instance and the state of the table
+ * @param props The properties to configure the table
+ * @returns The table instance and the state of the table
  */
+
 const useReactDataTable = <TData, TFilter extends FilterModel = Record<string, never>>(
   props: useReactDataTableProps<TData, TFilter>,
 ): useReactDataTableResult<TData, TFilter> => {
@@ -96,46 +99,48 @@ const useReactDataTable = <TData, TFilter extends FilterModel = Record<string, n
   const effectiveOnColumnPinningChange = onColumnPinningChange ?? setColumnPinningInternal;
 
   // If we active the manual filtering, we have to unset the filter function, else it still does automatic filtering
-  if (manualFiltering) columns.forEach((x) => (x.filterFn = undefined));
+  if (manualFiltering) for (const x of columns) x.filterFn = undefined;
 
   const internalColumns = columns.filter((x) => x.meta?.isHidden !== true);
   const skeletonColumns = internalColumns.map((column) => ({
     ...column,
     cell: () => <Skeleton />,
   }));
-  const skeletonData = Array.from({ length: paginationInternal.pageSize }, () => ({} as TData));
+  const skeletonData = Array.from({ length: paginationInternal.pageSize }, () => ({}) as TData);
 
   const columnFilters = useMemo(() => getColumnFilterFromModel(effectiveColumnFilters), [effectiveColumnFilters]);
   const sorting = useMemo(() => getSortingStateFromModel(effectiveSorting), [effectiveSorting]);
 
+  // known issue for tanstack with open ticket: https://github.com/facebook/react/issues/33057
+  // eslint-disable-next-line react-hooks/incompatible-library
   const table = useReactTable<TData>({
     data: isLoading ? skeletonData : data,
     columns: isLoading ? skeletonColumns : internalColumns,
 
     onColumnFiltersChange: (filtersOrUpdaterFn) => {
-      const newFilter = typeof filtersOrUpdaterFn !== "function" ? filtersOrUpdaterFn : filtersOrUpdaterFn(columnFilters);
+      const newFilter = typeof filtersOrUpdaterFn === "function" ? filtersOrUpdaterFn(columnFilters) : filtersOrUpdaterFn;
       return effectiveOnColumnFiltersChange(getModelFromColumnFilter(newFilter));
     },
     onPaginationChange: (paginationOrUpdaterFn) => {
-      const newFilter = typeof paginationOrUpdaterFn !== "function" ? paginationOrUpdaterFn : paginationOrUpdaterFn(effectivePagination);
+      const newFilter = typeof paginationOrUpdaterFn === "function" ? paginationOrUpdaterFn(effectivePagination) : paginationOrUpdaterFn;
       return effectiveOnPaginationChange(newFilter);
     },
     onSortingChange: (sortingOrUpdaterFn) => {
-      const newFilter = typeof sortingOrUpdaterFn !== "function" ? sortingOrUpdaterFn : sortingOrUpdaterFn(sorting);
+      const newFilter = typeof sortingOrUpdaterFn === "function" ? sortingOrUpdaterFn(sorting) : sortingOrUpdaterFn;
       return effectiveOnSortingChange(getModelFromSortingState(newFilter));
     },
     onRowSelectionChange: (rowSelectionOrUpdaterFn) => {
       const newRowSelection =
-        typeof rowSelectionOrUpdaterFn !== "function" ? rowSelectionOrUpdaterFn : rowSelectionOrUpdaterFn(effectiveRowSelection);
+        typeof rowSelectionOrUpdaterFn === "function" ? rowSelectionOrUpdaterFn(effectiveRowSelection) : rowSelectionOrUpdaterFn;
       return effectiveOnRowSelectionChange(newRowSelection);
     },
     onExpandedChange: (expandedOrUpdaterFn) => {
-      const newExpanded = typeof expandedOrUpdaterFn !== "function" ? expandedOrUpdaterFn : expandedOrUpdaterFn(effectiveExpanded);
+      const newExpanded = typeof expandedOrUpdaterFn === "function" ? expandedOrUpdaterFn(effectiveExpanded) : expandedOrUpdaterFn;
       return effectiveOnExpandedChange(newExpanded);
     },
     onColumnPinningChange: (columnPinningOrUpdaterFn) => {
       const newColumnPinning =
-        typeof columnPinningOrUpdaterFn !== "function" ? columnPinningOrUpdaterFn : columnPinningOrUpdaterFn(effectiveColumnPinning);
+        typeof columnPinningOrUpdaterFn === "function" ? columnPinningOrUpdaterFn(effectiveColumnPinning) : columnPinningOrUpdaterFn;
       return effectiveOnColumnPinningChange(newColumnPinning);
     },
 
@@ -184,7 +189,7 @@ const useReactDataTable = <TData, TFilter extends FilterModel = Record<string, n
 
   return {
     table,
-    columnFilters: effectiveColumnFilters as TFilter,
+    columnFilters: effectiveColumnFilters,
     pagination: effectivePagination,
     sorting: effectiveSorting,
     rowSelection: effectiveRowSelection,
@@ -199,4 +204,4 @@ const useReactDataTable = <TData, TFilter extends FilterModel = Record<string, n
   };
 };
 
-export { useReactDataTable, useReactDataTableProps, useReactDataTableResult };
+export { useReactDataTable };