+++ src/lib/utils/getColumnFilterFromModel.ts
@@ -2,6 +2,6 @@
 import { FilterModel } from "../types/TableState";
 
 const getColumnFilterFromModel = <TFilter extends FilterModel>(filter: TFilter): ColumnFiltersState =>
-  Object.entries(filter).map(([id, value]) => ({ id, value }));
+  Object.entries(filter).map(([id, value]) => ({ id, value: value as unknown }));
 
 export { getColumnFilterFromModel };