+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.15.0] - 2026-01-20
+
 ### Added
 
 - added the `usePersistentReactDataTableState` hook which saves it's own state into the local storage
@@ -456,7 +458,8 @@ columnHelper.display({
 
 - created package :tada:
 
-[unreleased]: https://github.com/neolution-ch/react-data-table/compare/5.14.0...HEAD
+[unreleased]: https://github.com/neolution-ch/react-data-table/compare/5.15.0...HEAD
+[5.15.0]: https://github.com/neolution-ch/react-data-table/compare/5.14.0...5.15.0
 [5.14.0]: https://github.com/neolution-ch/react-data-table/compare/5.13.1...5.14.0
 [5.13.1]: https://github.com/neolution-ch/react-data-table/compare/5.13.0...5.13.1
 [5.13.0]: https://github.com/neolution-ch/react-data-table/compare/5.12.0...5.13.0